rename login and register modal and move them in a folder
