+++ frontendV2/src/Router.tsx
@@ -4,6 +4,7 @@ import DemoScreen from '@pages/DemoScreen';
 import FAQScreen from '@pages/FAQScreen';
 import HomeScreen from '@pages/HomeScreen';
 import NewProfile from '@components/NewProfile';
+import Profiles from '@pages/Profiles';
 
 export default function Router() {
   return (
@@ -11,7 +12,7 @@ export default function Router() {
       <Route index path="/" element={<HomeScreen />} />
       <Route path="/demo" element={<DemoScreen />} />
       <Route path="/faq" element={<FAQScreen />} />
-      <Route path="/profiles" element={<NewProfile />} />
+      <Route path="/profiles" element={<Profiles />} />
     </Routes>
   );
 }
