+++ frontendV2/src/pages/Profiles.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Profiles() {
+  return <div>Profiles</div>;
+}
+
+export default Profiles;
